cx_911_agent/requirements.txt
google-cloud-aiplatform[adk,agent_engines]==1.130.0